Move ProcessImpl back to use_process_tree
